pyproject.toml
[tool.poetry]
name = "autotrading"
version = "0.1.0"
description = ""
authors = ["cjpark-data <cj.park@socialinvestinglab.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.23"
jwt = "^1.3.1"
requests = "^2.31.0"
pyupbit = "^0.2.33"
prophet = "^1.1.5"
python-dotenv = "^1.0.0"
yfinance = "^0.2.33"
statsmodels = "^0.14.1"
fredapi = "^0.5.1"
pandas = "^2.1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
